Add checkout screen

+++ app/src/main/java/com/example/gymappandroid/data/models/Product.kt
@@ -8,6 +8,6 @@ data class Product(
     val model: String = "",
     val description: String = "",
     val image: List<String> = listOf(),
-    val price: Double = 0.00,
+    val price: Long = 0,
     val availableSize: List<SizeOption> = listOf()
 ) : Serializable
