+++ app/src/main/java/com/example/gymappandroid/data/models/ShoppingCartItem.kt
@@ -1,11 +1,11 @@
 package com.example.gymappandroid.data.models
 
 data class ShoppingCartItem(
-    val shoppingCartItemId:String="",
+    val shoppingCartItemId: String = "",
     val itemId: String = "",
     val userId: String = "",
     val model: String = "",
     val imageSource: String = "",
-    val price: Double = 0.00,
+    val price: Long = 0,
     val specification: String = ""
 )
